Exams API

+++ api/def/exam.js
@@ -8,6 +8,26 @@ let exam_model = require('../../model/exam.js');
 let ExamDb = require('../../mock/mockedExam');
 
 
+
+router_exam.get('/', async function (req, res, next) {
+    let userId = req.uid;
+
+    if (apiUtility.validateParamsUndefined(userId)) //checking if the paramiters are defined
+        res.status(400).json(errors.PARAMS_UNDEFINED);
+
+    if (!apiUtility.validateParamsNumber(userId)) //checking if parameters are number, otherwise they are not valid
+        res.status(400).json(errors.PARAMS_WRONG_TYPE);
+    try {
+        let exam = await examimpl.getExam(examId);
+        if (exam === undefined)
+            res.status(404).send(errors.ENTITY_NOT_FOUND);
+        else
+            res.status(200).json(exam);
+    } catch (err) {
+        next(err);
+    }
+});
+
 router_exam.get('/:id', async function (req, res, next) {
     let examId = parseInt(req.params.id); // parsing the examID from the URL
     let userId = req.uid;
@@ -18,6 +38,14 @@ router_exam.get('/:id', async function (req, res, next) {
     if (!apiUtility.validateParamsNumber(userId, examId)) //checking if parameters are number, otherwise they are not valid
         res.status(400).json(errors.PARAMS_WRONG_TYPE);
     try {
+        let examRet = await examimpl.getExam(examId);
+        if (examRet === undefined)
+            return res.status(404).json(errors.ENTITY_NOT_FOUND);
+
+        let isOwner = await examimpl.validateOwner(userId, examId);
+        if (isOwner === false)
+            return res.status(401).json(errors.ACCESS_NOT_GRANTED);
+
         let exam = await examimpl.getExam(examId);
         if (exam === undefined)
             res.status(404).send(errors.ENTITY_NOT_FOUND);
@@ -71,7 +99,6 @@ router_exam.put('/:id', async function (req, res, next) {
     }
 });
 
-/*
 
 router_exam.delete('/:examId', async function (req, res, next) {
     let examId = parseInt(req.params.examId);
@@ -79,19 +106,27 @@ router_exam.delete('/:examId', async function (req, res, next) {
     let userId = req.uid;
 
     if (apiUtility.validateParamsUndefined(userId, examId))
-        res.status(400).json(errors.PARAMS_UNDEFINED);
+        return res.status(400).json(errors.PARAMS_UNDEFINED);
 
     if (!apiUtility.validateParamsNumber(userId, examId))
-        res.status(400).json(errors.PARAMS_WRONG_TYPE);
-
+        return res.status(400).json(errors.PARAMS_WRONG_TYPE);
+    
     try {
+        let examRet = await examimpl.getExam(examId);
+        if (examRet === undefined)
+            return res.status(404).json(errors.ENTITY_NOT_FOUND);
+
+        let isOwner = await examimpl.validateOwner(userId, examId);
+        if (isOwner === false)
+            return res.status(401).json(errors.ACCESS_NOT_GRANTED);
+
         await examimpl.deleteExam(userId, examId);
-        res.status(204).end();
+        return res.status(204).end();
     }
     catch (err) {
-        next(err);
+        return next(err);
     }
 });
 
-*/
+
 module.exports = router_exam;
